General commit for a reworking of the Species reference state thermo calculation. This is a reclarification of the reference state thermo calculations for individual species, and an expansion to handle liquid phase thermo needs.
There is now a virtual base class for the calculation of reference state
thermo functions for  individual species. It is called
SpeciesThermoInterpType.

There is also a class which allows for a complete general
calculation of the reference state species thermo for a phase,
GeneralSpeciesThermo.

Some of this new functionality may be relegated to ifdef blocks
in the future to limit the amount of code for users who don't
need the new functionality.