Added several missing thermodynamics routines to SurfPhase object.
Added error checking to make sure that a ThermoPhase object has
at least one species defined in it. Currently, it just won't work
if there are 0 species in the object, so we might as well throw
an error early to indicate this.

Added error checking to the SurfPhase object to indicate when
setCoverages is called with no coverages specified. This condition had
caused NaN's. Now, it throws a CanteraError.