Changed the algorithm for picking the component basis.
  It used to be that the components were picked based on the magnitude
of their mole numbers. I changed this to the magnitude of the
mole numbers multiplied by the # elements in the species.

  On one problem, a heptane combustion mechanism, this reduced  the
number of iterations in the VCS algorithm from 1080 to 40. Obviously,
the convergence rate within the VCS method
depends on the ID of the components. This change should be considered
as a trial balloon. It may fall apart on some other problem.

 I may be forestalling the obvious here. The more complete solution
is to formualte the hessian of the rxns involving the
major species, and then invert it, using cg probably, to get a better
iteration direction.