Bug fix / clarification to importPhase()
    Created a list of species early on in the new routine,
 formSpeciesXMLNodeList(). This list is the actual list of
species that will make up the phase. This is done before the
id of the SPeciesThermo or VPSSMgr is chosen, so that this id
is chosen based on just the species to be included in the
phase and not on all of the species that are in the database.

The overall amount of code is reduced, and the code is clarified.

Fixed an error in 	VPSSMgr_Water_HKFT.