Cleaning up erroneous 'sanity check' in RedlichKwongMFTP:pressure()
I'm not really sure what these lines of code were ever supposed
to achieve, but the effect is that you can never change the
object's thermodynamic state in a way that changes the pressure,
unless you are setting the pressure directly.  I'm leaving it
commented, in case we find at a later time that it serves some
purpose.