Correcting error in constDensityThermo:standardConcentration.
StandardConcentration(size_t k) should return the molarDensity
for species k at T and P with X_k = 1.  Because ConstDensityThermo
specifies a constant mass density (kg m^-3), the molarDensity
of the phase changes with composition.  This method previously
returned molarDensity() as the standardConcentration of species k.
This leads to a standard concentration that is a function of the
current composition, which is clearly problematic (i.e., incorrect).

Updated so that ConstDensityThermo:standardConcentration returns
density divided by molecularWeight of species k.