speciesIndex() takes atomic composition map
Phase::speciesIndex() can take an atomic composition map to find the
index of species with that composition. Due to the possibility of
isomers, the overloaded function always returns a vector<size_t>, even
if no isomers are found. Similarly, in the Python interface, passing a
dictionary to species_index returns a list even if only a single
species is found.

Resolves #603