[Thermo] Check that special species was specified
In the IonsFromNeutral phase, if a special-species is not included the
index is never set after initialization and remains at npos. The check
here ensures that a special-species is specified.

Also move building the elemVectorI outside of the jNeut loop. It doesn't
appear to need to be done for every neutral species, so this would save
some effort