[ctml2yaml] Add command line arguments
Add parsing of arguments passed when the script is run from the command
line. A new function is introduced to parse the arguments and is the
default function when the file is executed as a script. Update the
documentation to add the new function.

Add a new argument to convert that takes a string containing the content
of a CTML file. This allows convert to be used from within the Python
interpreter more easily.