[Thermo] Fix pure fluid state restoration after property calculation
After calculating properties using finite differences (e.g. cp), the fluid was
returned to its original state by setting the state in terms of pressure and
temperature. However, because setting the state using these variables is an
iterative calculation with some error tolerance, the state was not restored
exactly. By restoring the state based on the initial temperature and density,
the restored state is identical to the initial state.