Fix finding Apple's clang compiler
The previous check was too specific for the output of --version and did
not work with Clang 11.0. I hope that Xcode being somewhere in the
output will be more stable than the output starting with Apple LLVM. I
checked the output of Clang installed with Homebrew and it did not
include Xcode anywhere in it.