[Equil] Fix use of element potential method with IdealSolidSolnPhase
The implementation of IdealSolidSolnPhase::setToEquilState incorrectly used
the input array as the element potentials rather than the species chemical
potentials. The modified implementation corresponds to that of the
IdealGasPhase class.

The previous implementation seems to generally have resulted in an exception
being raised, which normally results in falling back to the Gibbs solver,
so this shouldn't have been producing incorrect results.