[Reactor/1D] Fix gas phase production on surfaces
The calculation for the production of gas phase species from
surface reactions assumed that the gas phase species were stored at the
beginning of the rate-of-production array. This is only the case when
the gas ThermoPhase instance is the first phase in the Kinetics thermo
vector, which is not guaranteed to be the case.

It happens that XML input usually forced the gas phase to be the first
phase in the thermo vector, so everything worked out fine. This was
because the associated phase(s) for a surface phase had to be listed in
an XML file, and the XML initializer reads that list to fill the
Kinetics thermo vector.

In the YAML format, the associated phase does not need to be explicitly
listed. Instead, instances of the appropriate phases are passed to the
constructor when the InterfaceKinetics is created. The gas phase can
then come before or after the surface phase, meaning there is no
guarantee of the order of species.

This change checks for the starting index of the gas phase species in
the InterfaceKinetics species-length arrays to make sure that the
appropriate rates of production are added for the gas phase species.

Fixes #902, #877, #873