[Cython] Add pure-Python import of 'interrupts'
Tools for creating standalone Python applications, such as PyInstaller,
can fail to pick up the dependence on the 'cantera.interrupts' module because
it is only imported and used from Cython. Explicitly importing this module
from another pure-Python module causes these tools to correctly see that the
interrupts module needs to be included in the application.

Fixes #914.