[CI] Fix missing Cython & NumPy dependencies
The Cython package for Python 3 is called cython3. We use python_cmd to
check for Cython now, which is Python 3.

The test comparisons are done with the Python that's running SCons,
which is Python 2 in this test. NumPy needs to be installed for that
Python to be able to do CSV file comparisons in the tests.