[CI] Remove Python 3.5 and 3.6 from macOS job
GitHub Actions is limited to 5 concurrent macOS runners. Having 10 macOS
jobs was causing a bottleneck in the CI runs. This change reduces the
number of concurrent jobs from a single run to 6, which will hopefully
be sustainable. The other versions of Python are currently tested on
Ubuntu and Windows as well.