[SCons/Python] Create the source distribution
Add a new option for the Python package to build a source distribution.
It is added as a new interface to simplify copying files. All of the
Cantera plus external library source code is copied into the interface
folder for building.

Actually building the source distribution is handled outside SCons by a
separate utility such as Setuptools or the pypa-build package. Building
is set up to use the PEP 517/518 style of builder with the
pyproject.toml file to configure the build backend and requirements.