[SCons/Python] Create the source distribution
Add a new option for the Python package to build a source distribution.
It is added as a new interface to simplify copying files. All of the
Cantera plus external library source code is copied into the interface
folder for building.

Actually building the source distribution is handled by the 'build'
package, which becomes a new dependency to build the sdist. 'build'
handles installing all the build-time dependencies in an isolated
environment using specifications in pyproject.toml.