[Thermo] Combine file-based and default constructors
For classes derived from ThermoPhase, instead of having separate
constructors for the default constructor and the constructor which
takes an input file and phase id, we can use the case where both
arguments are the default (empty string) to construct a phase without
using an input file. This eliminates the need to repeat any
initialization that takes place in the constructor.