Fixes #1001
Makes the `Reaction.listFromFile` example valid by adding the required
Kinetics object as the second argument (`gas`, created in the example
above.)

I'm not sure if this is the best solution: it has the major downside of
requiring `gas` to already be initialized to a `Kinetics` object with
the appropriate species. I'm new to Cantera, but I couldn't find an easy
way to make a "dummy" Kinetics object. Is there a better way to do this?

Maybe the example should just be removed?