Rename cleaning functions and add loggers
The loggers added here will eventually be used to produce all the output
from the build process that we control (that is, it won't replace
anything that SCons prints). The motivation for this change is to have
better configurability of when and where output is sent to the terminal,
for instance, this allows us to easily change to sending error messages
to stderr instead of stdout. Another thing this makes possible is builds
with varying levels of output, for instance, if a user only wants to see
error messages.