[Input] Set rate_units for reactions whenever possible
Setting the rate_units for a Reaction requires a Kinetics object. When created from
a YAML file, this is available while the Reaction is being constructed. For Reactions
created from CTI/XML or directly in code, this can be done when the Reaction is added
to a Kinetics object. Serializing using a non-default unit system is not possible if
neither of these conditions are met.