Enforce Python 3.6 or higher for building
Exit immediately if Python 2 is running SCons

Going forward, we want to use Python 3 syntax features, which cause
SConstruct to raise a SyntaxError during parsing in Python 2 and no code
is executed. As such, it is not possible to use the standard exception
flow to raise a useful error message unless we force SConstruct to use
only Python 2 compatible syntax. Therefore, this change introduces an
"f"-string that will raise a SyntaxError in Python 2 with a message
about running SCons with Python 3.