Refactor the loggers to a single instance
Set up a filter and output formatter so that there is only a single
instance of a logger class. Whether or not the level name is printed
with the message is determined by the print_level key in the extra
dictionary argument. This change clarifies the purpose of the previous
two logger instances. The output of error level and higher is sent to
the stderr stream, all lower levels are sent to stdout.