[Test] Avoid unnecessary creation of Transport objects
The default phase definition for gri30.yaml and h2o2.* includes a
transport model, and setting this slows down tests that don't need
transport properties.

For any tests changed, also switch to using the YAML versions of these
input files. The addition of N2 to h2o2.yaml, along with one reaction,
means that some indices had to change. This also uncovered a bug in
the Gibbs equilibrium solver, which is documented in #1023.