[1D] Do not impose bounds on inactive solution components
For flame simulations that don't use the 'eField' solution component,
replace the default solution bounds which require this component to
remain exactly zero with arbitrarily wide bounds. Small rounding
errors can occasionally cause the solver to want to have this
component take on a small nonzero value, which could not be satisfied
for any step size, resulting in solver failure.