[SCons] Disable OMP on XCode Clang
According to several sources online, XCode does not support OMP. This
disables checking for OMP support when Apple's clang (XCode) is being
used.

In addition, the library associated with GCC is called gomp, with LLVM
is called just omp, and with the Intel compiler is called iomp5. SCons
can check for multiple libraries sequentially, stopping when it finds
the first successful case. Due to this, the order of the checks are
important. Presumably, iomp5 and omp will only be installed for their
respective compilers, whereas gomp is likely to be installed for all the
compilers. Thus, gomp must go last in the order. This fixes the CI to
run the OpenMP sample correctly.