[sdist] Fix platform-dependent sdist
The config.h file for Cantera needs to have certain values set on the
platform that will build/install Cantera, not the platform building the
sdist. These values are now added to the config.h.in template by
setup.py. Values that are not platform dependent are still handled by
SConscript when the sdist is generated.