[sdist] Reduce wheel size on Linux
Setuptools includes CFLAGS from when Python was built, which may include
the flag to add debug symbols. This dramatically inflates the size of
our wheels, so this change disables those symbols. We also exclude
source files from the built wheel. _cantera.cpp ends up being ~10 MB
uncompressed, so this is a signifcant savings as well. Overall, the
wheel goes from ~100 MB to ~5 MB.