[Python] Fix ATOL in advance_to_steady_state
The docstring says that if undefined it should match the solver ATOL,
which makes sense to me. What the code in fact did was atol = self.rtol,
which I presume is a typo, so have changed to atol = self.atol.

Since RTOL is often ten or more orders of magnitude larger than ATOL,
this could make some significant differences to people trying to use this 
method. 

However, the changes could be bad: if now we can't ever reach the steady state,
or it takes much longer to reach steady state (already a complaint I hear),
folks may complain that this commit just made it a bunch slower and less stable. 

(But at least they're now getting what they ask for, instead of just getting
the wrong answer quicker).