Ensure registration of extension rates in host Cantera library
If a user application is linked statically to the Cantera library,
ExtensibleRate objects need to be registered in this copy of the Cantera
library rather than the one that is embedded in the Python module.
This is achieved by accessing the ReactionRateFactory from the main
application rather than from the Python module.