[CI] Use correct Python version on macOS
Linking to libpython with GitHub's Python 3.8 and 3.10 requires linking
to libintl, but no compatible version of this library is present in
the GitHub actions images (the Homebrew one has the wrong "macOS deployment
target"). This problem is fixed in GitHub's build of Python 3.11.

Scons 4.4.0 (and maybe other versions) can't be pip installed for the
Homebrew Python, since it tries to write man pages into an incorrect,
read-only directory.

The Homebrew-installed Python was never actually being used. All macOS
builds are done using Python cached in the GitHub actions image.