[Doc] Fix building Sphinx docs on macOS
macOS doesn't allow passing DYLD_LIBRARY_PATH through a shell, so
we need to use subprocess.call to be able to call sphinx-build while
including build/lib in the library path. The need to have the library
path set for this is a result of the transition to linking the Python
module against a shared version of libcantera.