Deprecate class IonsFromNeutralVPSSTP and class PDSS_IonsFromNeutral
This model has had a request for tests/examples out for many years (#267)
with no response, and the addition of consistency tests revealed multiple
errors in the implementation (documented in #1322). Further, there is no
mathematical description of the model or references to literature to use
as a reference in trying to correct any of these issues.

Resolves #1322.