[Doc] Revise species thermo documentation
Expand the description to cover both the calculation of standard-state
cp/h/s and the models that provide parameters affecting the equation
of state.

Also try to provide a description of the difference between the typical
definition of standard state and the distinction that Cantera sometimes
tries to make between "reference state" and "standard state" properties.