[sdist] Simplify extension setup to resolve duplicate cythonizing
When building the sdist via scons, the cython step was running twice because of the conditional. In addition, we were getting a warning about building the extension from multiple source files. These changes resolve both problems by cythonizing once if the files aren't present and only including the C++ sources in the extension.