[SCons] Replace CT_SUNDIALS_VERSION constant
The CT_SUNDIALS_VERSION constant was originally defined when SUNDIALS
only defined a string version number, so we couldn't do comparisons.
Since SUNDIALS v3.0, the version parts have been available as integers,
enabling comparison with the SUNDIALS-defined constants. This change
simplifies building the libraries because it removes the custom constant
and simplifies building the sdist/wheel because its one fewer constant
that has to be templated in.