[SCons] Refactor system SUNDIALS checks
Since we can run the C preprocessor to find version information about
Sundials, we can simplify many of the checks we have when the system
Sundials is specified. In particular, we can determine ahead of time
which calls we expect to find and which libraries to use, without having
to compile each one. In addition, the Sundials check is refactored into
buildutils to simplify SConstruct